Contact form

+++ db/migrate/20221223015333_create_quotes.rb
@@ -2,7 +2,7 @@ class CreateQuotes < ActiveRecord::Migration[7.0]
   def change
     create_table :quotes do |t|
       t.integer :status
-      t.float :amount
+      t.integer :amount
       t.text :comment
       t.references :vehicle, null: false, foreign_key: true
       t.references :freight, null: false, foreign_key: true
